Rename EmailAlgebra -> EmailPort
